requirements.txt
python-telegram-bot==21.4
SQLAlchemy==2.0.32
alembic==1.13.2
python-dotenv==1.0.1
httpx==0.27.0
APScheduler==3.10.4
loguru==0.7.2
